Make mdsh fail if any command fails & record errors
We capture the stderr of all commands and collect the Outputs of all
commands that failed. If any fails, we let mdsh fail with status 1 and
print the status and stderr of all failing commands to stderr.

This way integration of mdsh into CI is made possible, letting it act
as a “doctest” for markdown files, checking the integrity of for
example tutorials.

Fixes https://github.com/zimbatm/mdsh/issues/11